# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateTruckRequest:
    properties:
      is_active:
        description: Active status
        example: true
        type: boolean
      plate_number:
        description: Vehicle registration plate
        example: B-1234-XYZ
        type: string
      status:
        description: 'Status: available, in_use, maintenance'
        example: available
        type: string
      truck_type:
        description: Type/model of truck
        example: Fuso Box
        type: string
    required:
    - plate_number
    - status
    - truck_type
    type: object
...
  model.UpdateTruckRequest:
    properties:
      is_active:
        description: Update active status (optional)
        example: false
        type: boolean
      plate_number:
        description: Update plate number (optional)
        example: B-1234-XYZ
        type: string
      status:
        description: Update status (optional)
        example: maintenance
        type: string
      truck_type:
        description: Update truck type (optional)
        example: Fuso Box
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@jalurberlian.id
    name: API Support
    url: http://www.jalurberlian.id/support
  description: Real-time fleet management and order tracking system for PT. Jalur
    Berlian Makassar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Jalur Berlian Fleet Management API
  version: "1.0"
paths:
...
  /admin/trucks:
    get:
      description: Get all trucks with pagination support
      parameters:
      - description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of trucks
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List all trucks
      tags:
      - Trucks
    post:
      consumes:
      - application/json
      description: Register a new truck in the fleet
      parameters:
      - description: 'Truck details: plate_number, truck_type, status, is_active'
        in: body
        name: truck
        required: true
        schema:
          $ref: '#/definitions/model.CreateTruckRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Truck created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new truck
      tags:
      - Trucks
  /admin/trucks/{id}:
    delete:
      description: Deactivate a truck from fleet
      parameters:
      - description: Truck ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Truck deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Truck not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete truck
      tags:
      - Trucks
    get:
      description: Retrieve a specific truck by ID
      parameters:
      - description: Truck ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Truck details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Truck not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get truck details
      tags:
      - Trucks
    patch:
      consumes:
      - application/json
      description: Update truck details with PATCH support. Only provide fields you
        want to update (all fields are optional).
      parameters:
      - description: Truck ID
        in: path
        name: id
        required: true
        type: integer
      - description: Truck fields to update (only include fields you want to change)
        in: body
        name: truck
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTruckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Truck updated successfully with actual data from database
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input or validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Truck not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Partially update truck information
      tags:
      - Trucks
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"