# internal/model/truck.go
package model

import "time"

// Truck merepresentasikan tabel 'trucks' di database
type Truck struct {
	ID          int       `db:"id" json:"id"` // Immutable: Auto-generated by database
	PlateNumber string    `db:"plate_number" json:"plate_number" example:"B-1234-XYZ"` // Vehicle registration plate
	TruckType   string    `db:"truck_type" json:"truck_type" example:"Fuso Box"` // Type/model of truck
	Status      string    `db:"status" json:"status" example:"available"` // Status: available, in_use, maintenance
	IsActive    bool      `db:"is_active" json:"is_active" example:"true"` // Active status
	CreatedAt   time.Time `db:"created_at" json:"created_at"` // Immutable: Auto-set by database
}

// CreateTruckRequest adalah request body untuk membuat truck baru (hanya user-input fields)
type CreateTruckRequest struct {
	PlateNumber string `json:"plate_number" example:"B-1234-XYZ" binding:"required"` // Vehicle registration plate
	TruckType   string `json:"truck_type" example:"Fuso Box" binding:"required"` // Type/model of truck
	Status      string `json:"status" example:"available" binding:"required"` // Status: available, in_use, maintenance
	IsActive    bool   `json:"is_active" example:"true"`                      // Active status
}

// UpdateTruckRequest adalah request body untuk update truck (partial update).
// Menggunakan pointers untuk optional fields sehingga bisa membedakan antara
// "field tidak dikirim" vs "field dikirim dengan value kosong/false".
type UpdateTruckRequest struct {
	PlateNumber *string `json:"plate_number" example:"B-1234-XYZ"` // Update plate number (optional)
	TruckType   *string `json:"truck_type" example:"Fuso Box"` // Update truck type (optional)
	Status      *string `json:"status" example:"maintenance"` // Update status (optional)
	IsActive    *bool   `json:"is_active" example:"false"` // Update active status (optional)
}